change project structure. folders have been moved to the src folder
