+++ src/data/config.py
@@ -16,3 +16,4 @@ load_dotenv()
 BOT_TOKEN = os.environ.get('BOT_TOKEN')
 ADMINS = os.environ.get('ADMINS').split(',')
 IP = os.environ.get('ip')
+
