+++ src/filters/__init__.py
@@ -1,6 +1,3 @@
-from aiogram import Dispatcher
-
-from loader import dp
 # from .is_admin import AdminFilter
