+++ src/handlers/users/help.py
@@ -10,5 +10,6 @@ async def bot_help(message: types.Message):
             "/start - Начать диалог",
             "/reg - Войти / Зарегистрироваться",
             "/help - Получить справку")
-    
+
     await message.answer("\n".join(text))
+
