+++ src/handlers/users/reg.py
@@ -1,13 +1,12 @@
 from aiogram import types
-from aiogram.dispatcher.filters.builtin import CommandStart, Command
+from aiogram.dispatcher.filters.builtin import Command
 
-from keyboards.default import keyboards_login
+from src.keyboards.default import keyboards_login
 from loader import dp
 
 
 @dp.message_handler(Command("reg"))
 async def bot_register_user(message: types.Message):
-    print(message.chat.id)
     await message.answer(f"Будет зарегистрирован новый пользователь - {message.from_user.full_name}!",
                          reply_markup=keyboards_login)
     # здесь нужно сделать инлайн кнопки
@@ -28,3 +27,5 @@ async def sign_in(message: types.Message):
 async def sign_in(message: types.Message):
     await message.answer(f"Существующий пользователь нажмите 'SIGN IN'\n"
                          f"Создать нового пользователя нажмите 'SIGN UP'")
+
+
