+++ src/handlers/users/start.py
@@ -9,3 +9,4 @@ from loader import dp
 async def bot_start(message: types.Message):
     await message.answer(f"Привет, {message.from_user.full_name}!",
                          reply_markup=ReplyKeyboardRemove())  # Убрать кнопки
+
