+++ src/middlewares/__init__.py
@@ -1,5 +1,3 @@
-from aiogram import Dispatcher
-
 from loader import dp
 from .throttling import ThrottlingMiddleware
